Move mocks & unit tests to the subdirectories
