+++ tests/Mock/PostRepositoryMock.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TinyIB\Tests;
+namespace TinyIB\Tests\Mock;
 
 use TinyIB\Model\PostInterface;
 use TinyIB\Repository\PostRepositoryInterface;
